+++ packages/components/CHANGELOG.md
@@ -39,6 +39,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### Fixed
 
+- `Multi Select`, `Select`: not correctly updating when options within `p-optgroup` change dynamically
+  ([#4278](https://github.com/porsche-design-system/porsche-design-system/pull/4279))
 - `Flyout`, `Modal`, `Sheet`: `overflow: clip` causing issues
   ([#4296](https://github.com/porsche-design-system/porsche-design-system/pull/4296))
 - `Link Pure`: focus border of slotted anchor